[fix]머지안됨 이슈 해결
