+++ types/more/policy.ts
@@ -1,0 +1,5 @@
+export interface Policy {
+   id: number;
+   title: string;
+   detail: string;
+}
